Add test to see if we can require the library without any issues and add npm test script
